src/home/blog_scripts/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
newsapi-python==0.2.7